pyproject.toml
[project]
name = "Gunnersaurus"
version = "0.1.0"
description = "Python Discord Bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4==4.10.0",
    "discord-py==2.2.2",
    "fastf1>=3.5.3",
    "requests==2.31.0",
]
